Area 3 Final GrayBoxing Update
I fixed the elephants foot (placed to the side) moved the hallway, added the 3rd 2nd floor ladder, and added ceilings where necessary. There was interesting space on the second floor exit( a room right before you go to the shortcut) that we could possibly use to recreate one of the realistic photos we have.